Rename defaults.py to states.py in chopper
